Start versioning gitlog2changelog
This bash script's purpose is to convert the 'git log' informations in a
ChangeLog format per version of the tool with respect to the GNU standards
such as informations on authors and date of changes. This script is generic
and could be used for other tools, applications, or others under the
conditions of the GNU General Public License version 2.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>